+++ docs/docs/resources/changelog.md
@@ -12,6 +12,10 @@ Components with the <wa-badge variant="warning" pill>Experimental</wa-badge> bad
 During the alpha period, things might break! We take breaking changes very seriously, but sometimes they're necessary to make the final product that much better. We appreciate your patience!
 :::
 
+## Next
+
+- Fixed a bug in `<wa-switch>` where it would not properly change its "checked" state when its property changed.
+
 ## 3.0.0-alpha.7
 
 - Renamed applied.css to webawesome.css